Split up antsibull_nox.sessions.utils.
