+++ src/antsibull_nox/sessions/extra_checks.py
@@ -16,6 +16,8 @@ import nox
 
 from .utils import (
     compose_description,
+)
+from .utils.scripts import (
     run_bare_script,
 )
